ISSUE #3466 - Separate logic and view from unhandled error component
