+++ frontend/src/v5/ui/controls/errorMessage/unhandledError/unhandledError.component.tsx
@@ -15,60 +15,24 @@
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 import { isNetworkError } from '@/v5/validation/errors.helpers';
-import axios from 'axios';
-import { useEffect, useState } from 'react';
 import { NetworkError } from './networkError/networkError.component';
 import { UnexpectedError } from '../unexpectedError/unexpectedError.component';
 
 type UnhandledErrorProps = {
+	error: any;
 	expectedErrorValidators?: Array<(err) => boolean>;
 	className?: string;
-	initialError?: any;
-	setError?: (error: any) => void;
 };
 
 export const UnhandledError = ({
 	expectedErrorValidators = [],
 	className,
-	initialError,
-	setError,
+	error,
 }: UnhandledErrorProps) => {
-	const [showNetworkError, setShowNetworkError] = useState(null);
-	const [showUnexpectedError, setShowUnexpectedError] = useState(null);
-	const [interceptor, setInterceptor] = useState(null);
-
 	const isExpectedError = (err) => expectedErrorValidators.some((test) => test(err));
 
-	const onSuccess = (res) => {
-		setError?.(null);
-		setShowNetworkError(null);
-		setShowUnexpectedError(null);
-		return res;
-	};
-
-	const onError = (err) => {
-		setError?.(err);
-		setShowNetworkError(isNetworkError(err));
-		setShowUnexpectedError(!isExpectedError(err));
-		return Promise.reject(err);
-	};
-
-	const onMount = () => {
-		setInterceptor(axios.interceptors.response.use(
-			onSuccess,
-			onError,
-		));
-	};
-
-	const onUnmount = () => axios.interceptors.request.eject(interceptor);
-
-	useEffect(() => {
-		onMount();
-		if (initialError) onError(initialError);
-		return onUnmount;
-	}, []);
-
-	if (showNetworkError) return (<NetworkError className={className} />);
-	if (showUnexpectedError) return (<UnexpectedError className={className} />);
+	if (!error) return (<></>);
+	if (isNetworkError(error)) return (<NetworkError className={className} />);
+	if (!isExpectedError(error)) return (<UnexpectedError className={className} />);
 	return (<></>);
 };
